reuse projection fct to parse genome paths file
